added docker env file

+++ scripts/kill_container.sh
@@ -1,7 +1,19 @@
 #!/bin/bash
 
 echo "Killing all active containers"
-docker kill $(docker ps -q)
 
-docker rm $(docker ps -a -q)
+if [ -z "$(docker ps -q)" ]
+then
+   docker kill $(docker ps -q)
+fi
+   
+
+if [ -z "$(docker ps -a -q)" ]
+then
+   docker rm $(docker ps -a -q)
+fi
+
+#docker kill $(docker ps -q)
+
+#docker rm $(docker ps -a -q)
